Refactor vector database implementation
