+++ service/router.py
@@ -6,9 +6,9 @@ from semantic_router.route import Route
 from models.document import BaseDocumentChunk
 from models.query import RequestPayload
 from service.embedding import get_encoder
-from service.vector_database import VectorService, get_vector_service
 from utils.logger import logger
 from utils.summarise import SUMMARY_SUFFIX
+from vectordbs import BaseVectorDatabase, get_vector_service
 
 
 def create_route_layer() -> RouteLayer:
@@ -29,7 +29,7 @@ def create_route_layer() -> RouteLayer:
 
 
 async def get_documents(
-    *, vector_service: VectorService, payload: RequestPayload
+    *, vector_service: BaseVectorDatabase, payload: RequestPayload
 ) -> list[BaseDocumentChunk]:
     chunks = await vector_service.query(input=payload.input, top_k=25)
 
@@ -47,14 +47,14 @@ async def query(payload: RequestPayload) -> list[BaseDocumentChunk]:
     encoder = get_encoder(encoder_type=payload.encoder)
 
     if decision == "summarize":
-        vector_service: VectorService = get_vector_service(
+        vector_service: BaseVectorDatabase = get_vector_service(
             index_name=f"{payload.index_name}{SUMMARY_SUFFIX}",
             credentials=payload.vector_database,
             encoder=encoder,
         )
         return await get_documents(vector_service=vector_service, payload=payload)
 
-    vector_service: VectorService = get_vector_service(
+    vector_service: BaseVectorDatabase = get_vector_service(
         index_name=payload.index_name,
         credentials=payload.vector_database,
         encoder=encoder,
